# pkg/cache: fix race in Add() and Evict()