--- a/plugin/pkg/cache/cache_test.go
+++ b/plugin/pkg/cache/cache_test.go
@@ -3,11 +3,22 @@ package cache
 import "testing"
 
 func TestCacheAddAndGet(t *testing.T) {
-	c := New(4)
+	const N = shardSize * 4
+	c := New(N)
 	c.Add(1, 1)
 
 	if _, found := c.Get(1); !found {
 		t.Fatal("Failed to find inserted record")
+	}
+
+	for i := 0; i < N; i++ {
+		c.Add(uint64(i), 1)
+	}
+	for i := 0; i < N; i++ {
+		c.Add(uint64(i), 1)
+		if c.Len() != N {
+			t.Fatal("A item was unnecessarily evicted from the cache")
+		}
 	}
 }
 
@@ -27,6 +38,19 @@ func TestCacheLen(t *testing.T) {
 	c.Add(2, 2)
 	if l := c.Len(); l != 2 {
 		t.Fatalf("Cache size should %d, got %d", 2, l)
+	}
+}
+
+func TestCacheSharding(t *testing.T) {
+	const N = shardSize
+	c := New(N)
+	for i := 0; i < N*2; i++ {
+		c.Add(uint64(i), 1)
+	}
+	for i, s := range c.shards {
+		if s.Len() == 0 {
+			t.Errorf("Failed to populate shard: %d", i)
+		}
 	}
 }