### 1. Why is this pull request needed and what does it do?

This fixes a race in Add() when the shard is at max capacity and the key
being added is already stored.  Previously, the shard would evict a
random value - when all it needed to do was replace an existing value.
There was a race in how Evict() picked which key to remove, which would
cause concurrent calls to Evict() to remove the same key.

Additionally, this commit removes a lot of the lock contention and a
race around Add() and Evict() by changing them to immediately hold the
write lock.  Previously, they would check conditions with the read lock
held and not re-check those conditions once the write lock was acquired
(this is a race).

As an added bonus the cache is now slightly faster:
```
benchmark                      old ns/op     new ns/op     delta
BenchmarkCache-12              59.1          48.6          -17.77%
BenchmarkShard-12              64.5          59.6          -7.60%
BenchmarkShard_Parallel-12     119           111           -6.72%
```

### 2. Which issues (if any) are related?

N/A

### 3. Which documentation changes (if any) need to be made?

N/A

### 4. Does this introduce a backward incompatible change or deprecation?

N/A